review(server): constant for compare field
