+++ packages/twenty-server/src/engine/metadata-modules/flat-field-metadata/constants/flat-field-metadata-jsonb-properties.constant.ts
@@ -1,0 +1,8 @@
+import { FlatFieldMetadata } from "src/engine/metadata-modules/flat-field-metadata/types/flat-field-metadata.type";
+
+export const FLAT_FIELD_METADATA_JSONB_PROPERTIES = [
+  'defaultValue',
+  'options',
+  'settings',
+  'standardOverrides',
+] as const satisfies (keyof FlatFieldMetadata)[];
