+++ packages/twenty-server/src/engine/metadata-modules/flat-field-metadata/constants/flat-field-metadata-properties-to-compare.constant.ts
@@ -1,0 +1,15 @@
+import { FlatFieldMetadata } from "src/engine/metadata-modules/flat-field-metadata/types/flat-field-metadata.type";
+
+export const FLAT_FIELD_METADATA_PROPERTIES_TO_COMPARE = [
+  'defaultValue',
+  'description',
+  'icon',
+  'isActive',
+  'isLabelSyncedWithName',
+  'isUnique',
+  'label',
+  'name',
+  'options',
+  'standardOverrides',
+  'settings',
+] as const satisfies (keyof FlatFieldMetadata)[];
