+++ packages/twenty-server/src/engine/metadata-modules/flat-field-metadata/constants/flat-field-metadata-relation-properties-to-compare.constant.ts
@@ -1,0 +1,7 @@
+import { FlatFieldMetadataPropertiesToCompare } from "src/engine/metadata-modules/flat-field-metadata/types/flat-field-metadata-properties-to-compare.type";
+
+export const FLAT_FIELD_METADATA_RELATION_PROPERTIES_TO_COMPARE = [
+  'label',
+  'description',
+  'isActive',
+] as const satisfies FlatFieldMetadataPropertiesToCompare[];
