+++ packages/twenty-server/src/engine/metadata-modules/flat-field-metadata/types/flat-field-metadata-relation-properties-to-compare.type.ts
@@ -1,0 +1,4 @@
+import { FLAT_FIELD_METADATA_RELATION_PROPERTIES_TO_COMPARE } from "src/engine/metadata-modules/flat-field-metadata/constants/flat-field-metadata-relation-properties-to-compare.constant";
+
+export type FlatFieldMetadataRelationPropertiesToCompare =
+  (typeof FLAT_FIELD_METADATA_RELATION_PROPERTIES_TO_COMPARE)[number];
